requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.6
Django==3.2.7
django-3-jet==1.0.8
django-4-jet @ https://github.com/Barukimang/django-jet/archive/dev.zip
django-ckeditor==6.1.0
django-cuser==2017.3.16
django-js-asset==1.2.2
django-modeltranslation==0.17.3
django-rosetta==0.9.7
django-tinymce==3.3.0
django-widget-tweaks==1.4.8
gunicorn==20.1.0
idna==3.2
Pillow==8.3.2
polib==1.1.1
psycopg2==2.9.1
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
uWSGI==2.0.19.1
